python/requirements.txt
certifi==2023.7.22
psycopg2-binary==2.9.6
psycopg[binary]==3.1.8
psycopg-pool==3.1.7
requests==2.31.0
python-pptx==0.6.21
fastapi==0.100.0
azure-functions==1.15.0
azure-storage-blob==12.17.0
azure-functions-durable
zstandard==0.21.0
uvicorn==0.23.1
pydantic==2.1.1
pytz==2023.3
pyarrow==12.0.1
numpy==1.24.3
pandas==2.1.0
geopandas==0.14.0
python-dotenv==1.0.0
PyYAML==6.0
scikit-learn==1.3.0
httpx
aiohttp==3.11.18